Fixes logging out user after updating email
I noticed that when you update a user's email, you get immediately logged out. I imagine this is because Passport seeds its login token with the user's email, and once that changes, it no longer recognizes the login token as valid. I found this discussion: 
https://stackoverflow.com/questions/24493243/update-logged-in-user-details-in-session
which says that you need to re-login to update the session.